requirements.txt
flask
redis
flask-SQLAlchemy
psycopg2
uWSGI
python-dotenv
